requirements.txt
Django==4.2.20
psycopg2-binary==2.9.10
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0